session security url copy paste issue resolved

+++ src/app/ministry-login/ministry-login.component.ts
@@ -19,7 +19,13 @@ export class MinistryLoginComponent extends AppComponent implements OnInit {
 
 
   ngOnInit(): void {
+    if(sessionStorage.getItem('userType')=="<ministry>" && sessionStorage.getItem('<ministryId>')!=null){
     sessionStorage.clear();
+    }
+    else{
+      sessionStorage.clear();
+      this.router.navigate(['<ministryLogin>']);
+    }
   }
   ministryUid:string;
   ministryPassword:string;
